Renamed package shell to shellutils
