+++ pkg/kubernetesutils/KubernetesControlplaneHost.go
@@ -8,7 +8,7 @@ import (
 	"github.com/asciich/asciichgolangpublic/parameteroptions"
 	"github.com/asciich/asciichgolangpublic/pkg/contextutils"
 	"github.com/asciich/asciichgolangpublic/pkg/logging"
-	"github.com/asciich/asciichgolangpublic/shell/shelllinehandler"
+	"github.com/asciich/asciichgolangpublic/shellutils/shelllinehandler"
 	"github.com/asciich/asciichgolangpublic/tracederrors"
 )
